Import Glass DTS (2/3)
The DTS import is split into 3 commits due to the glass dts spanning 3
soc dts files. It is split so future upstreaming is easier.

The first commit is QC605 which is the primary soc for glass.
This commit is SDM670, which seems to have some similar changes to qc605
The third commit is sxr1120 which has an unreleated battery change and
imports for qc605's audio dts changed to the new glass one